Add files via upload
Hetionet parsing to obtain nodes_sideeffect.csv, which contains umls_cui  and their corresponding names. A UMLS CUI (Concept Unique Identifier) is a unique code assigned by the Unified Medical Language System (UMLS) to a specific biomedical concept to link it across different vocabularies and terminologies. For example, the CUI for "naproxen" connects its representation in RxNorm to its representation in SNOMED CT. A CUI ensures that different terms with the same meaning, like "kidney" and "renal," are all linked to a single, unified concept.